Add tests, minor fixes.

+++ implementations/simplewasm/src/index.c
@@ -4,10 +4,10 @@
 #define TRUE  (1==1)
 
 EMSCRIPTEN_KEEPALIVE
-unsigned long long checkPrime(unsigned long long num) {
+short checkPrime(unsigned long long num) {
     if (num == 2ull || num == 3ull) return TRUE;
     if (num <= 1ull || num % 2ull == 0 || num % 3ull == 0 ) return FALSE;  
-    for (long i = 5ull; i * i <= num ; i+=6ull) {
+    for (unsigned long long i = 5ull; i * i <= num ; i+=6ull) {
         if (num % i == 0ull || num % (i + 2ull) == 0) {
           return FALSE;
         }
